refactor: move card components out of views folder
