+++ src/components/TeacherComponents/SpecialistSubjectListingCardSection/SpecialistSubjectListingCardSection.stories.tsx
@@ -1,0 +1,20 @@
+import type { Meta, StoryObj } from "@storybook/react";
+
+import { specialistSubjectListingFixture } from "../../TeacherViews/SpecialistSubjectListing/SpecialistSubjectListing.fixture";
+
+import Component from "./SpecialistSubjectCardSection";
+
+const meta: Meta<typeof Component> = {
+  component: Component,
+};
+
+export default meta;
+type Story = StoryObj<typeof Component>;
+
+export const SpecialistUnitListingPage: Story = {
+  args: {
+    heading: "Specialist",
+    summary: "This is a summary of the specialist subjects.",
+    subjects: specialistSubjectListingFixture,
+  },
+};
